Fix use of `raise_error` matcher without providing a specific error warnings
When using `raise_error` without a specific exception or exception
message, any exception may cause the test to pass, even a
`NoMethodError` due to a typo, so rspec rightfully warns about doing it.

I've converted the tests which use a trivial "raises an error"
description to an empty one, which will cause rspec to automatically
generate a description for based on the assert being done (for instance
"should raise IndexError").