Fix `instance variable @size not initialized` warning
This warning happened because whenever a `Cons` is created via
`Hamster::List.from_enum` the `Cons#initialize` method is never run, and
thus `@size` is not set to `nil`, which then later causes a warning when
`Cons#cached_size?` is called.

As a fix, we now use `||=` to make sure that `@size` is initialized in
that method before we read it.