Fix `instance variable @expected not initialized` warning
One of the test cases did not use an `@expected` but instead hardcoded
the test so the test description came up as `returns nil`.

Because the tests description in this case can be trivially generated
from the expect itself, let's just remove all descriptions and let
RSpec automatically generate them.