Fix `too many arguments for format string` warning
When testing `Vector#rotate` with a `String`, before calling the
underlying `Array#rotate` (which is where the `TypeError` comes from) we
modulo the argument with the size, which would generate a "too many
arguments for format string warning" because String also implements `%`,
but obviously not the right one; luckily this still would work and so
the testcase would pass.

Unfortunately, besides the warning this actually hid a small bug where
we would actually return a `NoMethodError` for non-numeric arguments
that do not implement `%`, instead of the expected `TypeError`.

To fix this:

* let's use the `#modulo` method instead, which is an alias for `%`
  that `String` does not implement
* let's handle any `NoMethodError` that comes from our `#modulo`
  attempt and wrap it with a `TypeError` for all arguments